gozen/rikugun_sanbou/shikan/hohei/__init__.py
"""
歩兵（hohei）- Gemini API 並列

士官の下で、検証作業を担当する。
Gemini APIを使用して並列で分析を実行する。
"""

import asyncio
import os
from datetime import datetime

# Gemini API（GCP Vertex AI）
try:
    from google.cloud import aiplatform
    GEMINI_AVAILABLE = True
except ImportError:
    GEMINI_AVAILABLE = False


class Hohei:
    """
    歩兵クラス
    
    役割：
    - 士官からの検証指示を実行
    - Gemini APIを使用した分析
    - レビュー・監査
    """
    
    def __init__(self, worker_id: int):
        self.role = "歩兵"
        self.worker_id = worker_id
        self.superior = "士官"
        
        # Gemini API初期化
        if GEMINI_AVAILABLE and os.getenv('GCP_PROJECT_ID'):
            self.gemini_enabled = True
        else:
            self.gemini_enabled = False
    
    async def execute(self, verification_task: dict) -> dict:
        """
        検証タスクを実行
        """
        print(f"[歩兵{self.worker_id}] 検証開始: {verification_task.get('name', 'N/A')}")
        
        # 検証タイプに応じた処理
        task_type = verification_task.get('type', 'general')
        
        if task_type == 'cost_analysis':
            analysis = await self._analyze_cost(verification_task)
        elif task_type == 'operational_load':
            analysis = await self._analyze_operational_load(verification_task)
        elif task_type == 'risk_analysis':
            analysis = await self._analyze_risk(verification_task)
        elif task_type == 'alternative_evaluation':
            analysis = await self._evaluate_alternatives(verification_task)
        else:
            analysis = {'type': 'general', 'result': 'OK'}
        
        result = {
            'worker_id': self.worker_id,
            'task_id': verification_task.get('id'),
            'status': 'completed',
            'analysis': analysis,
            'timestamp': datetime.now().isoformat()
        }
        
        print(f"[歩兵{self.worker_id}] 検証完了")
        return result
    
    async def _analyze_cost(self, task: dict) -> dict:
        """コスト分析"""
        return {
            'type': 'cost_analysis',
            'initial_cost': '¥7,000〜¥20,000',
            'monthly_cost': '¥5,000〜¥15,000',
            'recommendation': '段階的投資を推奨'
        }
    
    async def _analyze_operational_load(self, task: dict) -> dict:
        """運用負荷分析"""
        return {
            'type': 'operational_load',
            'single_operator': True,
            'estimated_hours_per_week': '2-4時間',
            'automation_potential': 'high',
            'recommendation': '自動化で負荷軽減可能'
        }
    
    async def _analyze_risk(self, task: dict) -> dict:
        """リスク分析"""
        return {
            'type': 'risk_analysis',
            'risks': [
                {'name': '過剰設計', 'severity': 'medium', 'mitigation': '段階的導入'},
                {'name': '学習曲線', 'severity': 'high', 'mitigation': 'ドキュメント整備'},
                {'name': 'コスト超過', 'severity': 'low', 'mitigation': '予算管理'}
            ]
        }
    
    async def _evaluate_alternatives(self, task: dict) -> dict:
        """代替案評価"""
        return {
            'type': 'alternative_evaluation',
            'alternatives': [
                {'name': 'Docker Compose', 'score': 8, 'reason': 'シンプル・低コスト'},
                {'name': 'k3s', 'score': 7, 'reason': 'スケーラブル・学習曲線急'},
                {'name': 'Kubernetes', 'score': 5, 'reason': '過剰・コスト高'}
            ],
            'recommendation': 'Docker Compose → k3s の段階的移行'
        }


# モジュールレベル関数
async def execute(worker_id: int, verification_task: dict) -> dict:
    """歩兵の実行（モジュールレベル関数）"""
    hohei = Hohei(worker_id)
    return await hohei.execute(verification_task)
