requirements.txt
# Project GOZEN - Python Dependencies

# Core
pyyaml>=6.0
anthropic>=0.40.0
google-generativeai>=0.8.0
asyncio-compat>=0.1.0

# Gemini API (Optional - for rikugun)
google-cloud-aiplatform>=1.38.0

# CLI
argparse

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0
